requirements.txt
astroid==2.15.6
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dill==0.3.7
Flask==2.3.2
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pycodestyle==2.11.0
pylint==2.17.5
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
tomlkit==0.12.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
